+++ pkg/sidecar/sidecar_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/spiffe/go-spiffe/v2/spiffeid"
 	"github.com/spiffe/go-spiffe/v2/svid/x509svid"
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
-	"github.com/spiffe/spiffe-helper/internal/spiffetest"
+	"github.com/spiffe/spiffe-helper/test/spiffetest"
 	"github.com/spiffe/spiffe-helper/test/util"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
